Use google style docstrings (napoleon), optional book theme (#64)
The google style docstrings provided by napoleon extension are easier to write and read than the traditional ones, so
use this by default. Activate viewsource as well. A popular theme is the book theme which has some nice features, adding option for that which we could later take as default if desired.